GlobSearch: Change return type (#20)
* GlobSearch: Always return a list, change param name

It's a lot nicer if we can always expect a single type to be returned. It makes handling returns a lot easier on any devs. This PR changes the return type to always be a list[Path]. It also greatly simplifies some basic logic when looping over the search results.

* Chapters: Fix globsearch return check

* Revert un-shadowing builtin `dir`