do_audio: fix audio processing when there's no extractor (#28)
In `vsmuxtools.do_audio()`, when passing an AudioNode the extractor is set to
`None`. <https://github.com/Jaded-Encoding-Thaumaturgy/vs-muxtools/blob/v0.2.0/vsmuxtools/extension/audio.py#L93-L95>

This worked fine in muxtools v0.1.0 but then v0.2.0 changed how this `audio`
variable was set, with it now only being set if an extractor was defined. So in
the AudioNode case the variable is unbound and then:

    Traceback (most recent call last):
      File "test.py", line 5, in <module>
        do_audio(node)
      File ".venv/Lib/site-packages/vsmuxtools/extension/audio.py", line 105, in do_audio
        return mt_audio(fileIn, track, trims, fps, num_frames, extractor, trimmer, encoder, quiet, output)
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      File "muxtools/functions.py", line 83, in do_audio
        if not isinstance(audio, AudioFile):
                          ^^^^^
    UnboundLocalError: cannot access local variable 'audio' where it is not associated with a value

Minimal repro:

    from vsmuxtools import do_audio
    from vstools import core

    node = core.std.BlankAudio()
    do_audio(node)

What v0.1.0 was doing was a `audio = ensure_path_exists(fileIn, do_audio)`, so
this patch brings that back for the no-extractor case.